Crate deranged

deranged is a proof-of-concept implementation of ranged integers.

Structs

OptionRangedI8

An optional RangedI8; similar to Option<RangedI8> with better optimization.

OptionRangedI16

An optional RangedI16; similar to Option<RangedI16> with better optimization.

OptionRangedI32

An optional RangedI32; similar to Option<RangedI32> with better optimization.

OptionRangedI64

An optional RangedI64; similar to Option<RangedI64> with better optimization.

OptionRangedI128

An optional RangedI128; similar to Option<RangedI128> with better optimization.

OptionRangedIsize

An optional RangedIsize; similar to Option<RangedIsize> with better optimization.

OptionRangedU8

An optional RangedU8; similar to Option<RangedU8> with better optimization.

OptionRangedU16

An optional RangedU16; similar to Option<RangedU16> with better optimization.

OptionRangedU32

An optional RangedU32; similar to Option<RangedU32> with better optimization.

OptionRangedU64

An optional RangedU64; similar to Option<RangedU64> with better optimization.

OptionRangedU128

An optional RangedU128; similar to Option<RangedU128> with better optimization.

OptionRangedUsize

An optional RangedUsize; similar to Option<RangedUsize> with better optimization.

ParseIntError

An error which can be returned when parsing an integer.

RangedI8

An i8 that is known to be in the range MIN..=MAX.

RangedI16

An i16 that is known to be in the range MIN..=MAX.

RangedI32

An i32 that is known to be in the range MIN..=MAX.

RangedI64

An i64 that is known to be in the range MIN..=MAX.

RangedI128

An i128 that is known to be in the range MIN..=MAX.

RangedIsize

An isize that is known to be in the range MIN..=MAX.

RangedU8

A u8 that is known to be in the range MIN..=MAX.

RangedU16

A u16 that is known to be in the range MIN..=MAX.

RangedU32

A u32 that is known to be in the range MIN..=MAX.

RangedU64

A u64 that is known to be in the range MIN..=MAX.

RangedU128

A u128 that is known to be in the range MIN..=MAX.

RangedUsize

A usize that is known to be in the range MIN..=MAX.

TryFromIntError

The error type returned when a checked integral type conversion fails.

Type Aliases

Option_ri8

Equivalent to OptionRangedI8

Option_ri16

Equivalent to OptionRangedI16

Option_ri32

Equivalent to OptionRangedI32

Option_ri64

Equivalent to OptionRangedI64

Option_ri128

Equivalent to OptionRangedI128

Option_risize

Equivalent to OptionRangedIsize

Option_ru8

Equivalent to OptionRangedU8

Option_ru16

Equivalent to OptionRangedU16

Option_ru32

Equivalent to OptionRangedU32

Option_ru64

Equivalent to OptionRangedU64

Option_ru128

Equivalent to OptionRangedU128

Option_rusize

Equivalent to OptionRangedUsize

ri8

Equivalent to RangedI8

ri16

Equivalent to RangedI16

ri32

Equivalent to RangedI32

ri64

Equivalent to RangedI64

ri128

Equivalent to RangedI128

risize

Equivalent to RangedIsize

ru8

Equivalent to RangedU8

ru16

Equivalent to RangedU16

ru32

Equivalent to RangedU32

ru64

Equivalent to RangedU64

ru128

Equivalent to RangedU128

rusize

Equivalent to RangedUsize