Struct deranged::RangedU16

pub struct RangedU16<const MIN: u16, const MAX: u16>(/* private fields */);

A u16 that is known to be in the range MIN..=MAX.

Implementations

impl RangedU16<0, 0>

pub const fn exact<const VALUE: u16>() -> RangedU16<VALUE, VALUE>

impl<const MIN: u16, const MAX: u16> RangedU16<MIN, MAX>

pub const MIN: Self = _

The smallest value that can be represented by this type.

pub const MAX: Self = _

The largest value that can be represented by this type.

pub const unsafe fn new_unchecked(value: u16) -> Self

Creates a ranged integer without checking the value.

Safety

The value must be within the range MIN..=MAX.

pub const fn get(self) -> u16

Returns the value as a primitive type.

pub const fn new(value: u16) -> Option<Self>

Creates a ranged integer if the given value is in the range MIN..=MAX.

pub const fn new_static<const VALUE: u16>() -> Self

Creates a ranged integer with a statically known value. Fails to compile if the value is not in range.

pub const fn new_saturating(value: u16) -> Self

Creates a ranged integer with the given value, saturating if it is out of range.

pub const fn expand<const NEW_MIN: u16, const NEW_MAX: u16>( self, ) -> RangedU16<NEW_MIN, NEW_MAX>

Expand the range that the value may be in. Fails to compile if the new range is not a superset of the current range.

pub const fn narrow<const NEW_MIN: u16, const NEW_MAX: u16>( self, ) -> Option<RangedU16<NEW_MIN, NEW_MAX>>

Attempt to narrow the range that the value may be in. Returns None if the value is outside the new range. Fails to compile if the new range is not a subset of the current range.

pub fn from_str_radix(src: &str, radix: u32) -> Result<Self, ParseIntError>

Converts a string slice in a given base to an integer.

The string is expected to be an optional + or - sign followed by digits. Leading and trailing whitespace represent an error. Digits are a subset of these characters, depending on radix:

0-9

a-z

A-Z

Panics

Panics if radix is not in the range 2..=36.

Examples

Basic usage:

assert_eq!(RangedU16::<5, 10>::from_str_radix("A", 16), Ok(RangedU16::new_static::<10>()));

pub const fn checked_add(self, rhs: u16) -> Option<Self>

Checked integer addition. Computes self + rhs, returning None if the resulting value is out of range.

pub const unsafe fn unchecked_add(self, rhs: u16) -> Self

Unchecked integer addition. Computes self + rhs, assuming that the result is in range.

Safety

The result of self + rhs must be in the range MIN..=MAX.

pub const fn checked_sub(self, rhs: u16) -> Option<Self>

Checked integer addition. Computes self - rhs, returning None if the resulting value is out of range.

pub const unsafe fn unchecked_sub(self, rhs: u16) -> Self

Unchecked integer subtraction. Computes self - rhs, assuming that the result is in range.

Safety

The result of self - rhs must be in the range MIN..=MAX.

pub const fn checked_mul(self, rhs: u16) -> Option<Self>

Checked integer addition. Computes self * rhs, returning None if the resulting value is out of range.

pub const unsafe fn unchecked_mul(self, rhs: u16) -> Self

Unchecked integer multiplication. Computes self * rhs, assuming that the result is in range.

Safety

The result of self * rhs must be in the range MIN..=MAX.

pub const fn checked_div(self, rhs: u16) -> Option<Self>

Checked integer addition. Computes self / rhs, returning None if rhs == 0 or if the resulting value is out of range.

pub const unsafe fn unchecked_div(self, rhs: u16) -> Self

Unchecked integer division. Computes self / rhs, assuming that rhs != 0 and that the result is in range.

Safety

self must not be zero and the result of self / rhs must be in the range MIN..=MAX.

pub const fn checked_div_euclid(self, rhs: u16) -> Option<Self>

Checked Euclidean division. Computes self.div_euclid(rhs), returning None if rhs == 0 or if the resulting value is out of range.

pub const unsafe fn unchecked_div_euclid(self, rhs: u16) -> Self

Unchecked Euclidean division. Computes self.div_euclid(rhs), assuming that rhs != 0 and that the result is in range.

Safety

self must not be zero and the result of self.div_euclid(rhs) must be in the range MIN..=MAX.

pub const fn rem<const RHS_VALUE: u16>( self, rhs: RangedU16<RHS_VALUE, RHS_VALUE>, ) -> RangedU16<0, RHS_VALUE>

Remainder. Computes self % rhs, statically guaranteeing that the returned value is in range.

pub const fn checked_rem(self, rhs: u16) -> Option<Self>

Checked integer remainder. Computes self % rhs, returning None if rhs == 0 or if the resulting value is out of range.

pub const unsafe fn unchecked_rem(self, rhs: u16) -> Self

Unchecked remainder. Computes self % rhs, assuming that rhs != 0 and that the result is in range.

Safety

self must not be zero and the result of self % rhs must be in the range MIN..=MAX.

pub const fn checked_rem_euclid(self, rhs: u16) -> Option<Self>

Checked Euclidean remainder. Computes self.rem_euclid(rhs), returning None if rhs == 0 or if the resulting value is out of range.

pub const unsafe fn unchecked_rem_euclid(self, rhs: u16) -> Self

Unchecked Euclidean remainder. Computes self.rem_euclid(rhs), assuming that rhs != 0 and that the result is in range.

Safety

self must not be zero and the result of self.rem_euclid(rhs) must be in the range MIN..=MAX.

pub const fn checked_neg(self) -> Option<Self>

Checked negation. Computes -self, returning None if the resulting value is out of range.

pub const unsafe fn unchecked_neg(self) -> Self

Unchecked negation. Computes -self, assuming that -self is in range.

Safety

The result of -self must be in the range MIN..=MAX.

pub const fn neg(self) -> Self

Negation. Computes self.neg(), failing to compile if the result is not guaranteed to be in range.

pub const fn checked_shl(self, rhs: u32) -> Option<Self>

Checked shift left. Computes self << rhs, returning None if the resulting value is out of range.

pub const unsafe fn unchecked_shl(self, rhs: u32) -> Self

Unchecked shift left. Computes self << rhs, assuming that the result is in range.

Safety

The result of self << rhs must be in the range MIN..=MAX.

pub const fn checked_shr(self, rhs: u32) -> Option<Self>

Checked shift right. Computes self >> rhs, returning None if the resulting value is out of range.

pub const unsafe fn unchecked_shr(self, rhs: u32) -> Self

Unchecked shift right. Computes self >> rhs, assuming that the result is in range.

Safety

The result of self >> rhs must be in the range MIN..=MAX.

pub const fn checked_pow(self, exp: u32) -> Option<Self>

Checked exponentiation. Computes self.pow(exp), returning None if the resulting value is out of range.

pub const unsafe fn unchecked_pow(self, exp: u32) -> Self

Unchecked exponentiation. Computes self.pow(exp), assuming that the result is in range.

Safety

The result of self.pow(exp) must be in the range MIN..=MAX.

pub const fn saturating_add(self, rhs: u16) -> Self

Saturating integer addition. Computes self + rhs, saturating at the numeric bounds.

pub const fn saturating_sub(self, rhs: u16) -> Self

Saturating integer subtraction. Computes self - rhs, saturating at the numeric bounds.

pub const fn saturating_mul(self, rhs: u16) -> Self

Saturating integer multiplication. Computes self * rhs, saturating at the numeric bounds.

pub const fn saturating_pow(self, exp: u32) -> Self

Saturating integer exponentiation. Computes self.pow(exp), saturating at the numeric bounds.

pub const fn wrapping_add(self, rhs: u16) -> Self

Wrapping integer addition. Computes self + rhs, wrapping around the numeric bounds.

pub const fn wrapping_sub(self, rhs: u16) -> Self

Wrapping integer subtraction. Computes self - rhs, wrapping around the numeric bounds.