Crate deranged

Structs

OptionRangedI8

A RangedI8 that is optional. Equivalent to Option<RangedI8> with niche value optimization.

OptionRangedI16

A RangedI16 that is optional. Equivalent to Option<RangedI16> with niche value optimization.

OptionRangedI32

A RangedI32 that is optional. Equivalent to Option<RangedI32> with niche value optimization.

OptionRangedI64

A RangedI64 that is optional. Equivalent to Option<RangedI64> with niche value optimization.

OptionRangedI128

A RangedI128 that is optional. Equivalent to Option<RangedI128> with niche value optimization.

OptionRangedIsize

A RangedIsize that is optional. Equivalent to Option<RangedIsize> with niche value optimization.

OptionRangedU8

A RangedU8 that is optional. Equivalent to Option<RangedU8> with niche value optimization.

OptionRangedU16

A RangedU16 that is optional. Equivalent to Option<RangedU16> with niche value optimization.

OptionRangedU32

A RangedU32 that is optional. Equivalent to Option<RangedU32> with niche value optimization.

OptionRangedU64

A RangedU64 that is optional. Equivalent to Option<RangedU64> with niche value optimization.

OptionRangedU128

A RangedU128 that is optional. Equivalent to Option<RangedU128> with niche value optimization.

OptionRangedUsize

A RangedUsize that is optional. Equivalent to Option<RangedUsize> with niche value optimization.

ParseIntError

RangedI8

An i8 that is known to be in the range MIN..=MAX.

RangedI16

An i16 that is known to be in the range MIN..=MAX.

RangedI32

An i32 that is known to be in the range MIN..=MAX.

RangedI64

An i64 that is known to be in the range MIN..=MAX.

RangedI128

An i128 that is known to be in the range MIN..=MAX.

RangedIsize

An isize that is known to be in the range MIN..=MAX.

RangedU8

A u8 that is known to be in the range MIN..=MAX.

RangedU16

A u16 that is known to be in the range MIN..=MAX.

RangedU32

A u32 that is known to be in the range MIN..=MAX.

RangedU64

A u64 that is known to be in the range MIN..=MAX.

RangedU128

A u128 that is known to be in the range MIN..=MAX.

RangedUsize

A usize that is known to be in the range MIN..=MAX.

TryFromIntError